Remove SlowMoDriveCommand and DriveUntilDistance classes; refactor drive commands in RobotContainer
